Refactor :hammer:
