+++ whisper_live/transcriber/transcriber_faster_whisper.py
